test: stabilize websocket sandbox status seeding
