+++ packages/control-plane/test/integration/websocket-sandbox.test.ts
@@ -1,28 +1,14 @@
 import { describe, it, expect } from "vitest";
-import { initNamedSession, openSandboxWs, seedSandboxAuth, queryDO } from "./helpers";
+import {
+  initNamedSession,
+  openSandboxWs,
+  seedSandboxAuth,
+  queryDO,
+  waitForSandboxStatus,
+} from "./helpers";
 
 const SANDBOX_TOKEN = "test-sandbox-auth-token-abc123";
 const SANDBOX_ID = "sb-integration-test";
-const DEFAULT_WAIT_FOR_SANDBOX_STATUS_TIMEOUT_MS = 3000;
-
-async function waitForSandboxStatus(
-  stub: DurableObjectStub,
-  status: string,
-  timeoutMs = DEFAULT_WAIT_FOR_SANDBOX_STATUS_TIMEOUT_MS
-): Promise<void> {
-  const deadline = Date.now() + timeoutMs;
-  let lastStatus: string | undefined;
-  while (Date.now() < deadline) {
-    const rows = await queryDO<{ status: string }>(stub, "SELECT status FROM sandbox");
-    lastStatus = rows[0]?.status;
-    if (lastStatus === status) return;
-    await new Promise((r) => setTimeout(r, 100));
-  }
-
-  throw new Error(
-    `Timed out after ${timeoutMs}ms waiting for sandbox status "${status}"; last status was "${lastStatus ?? "missing"}"`
-  );
-}
 
 describe("Sandbox WebSocket (via SELF.fetch)", () => {
   it("upgrade with valid auth returns 101", async () => {
